+++ docs/requirements.txt
@@ -1,0 +1,3 @@
+myst-parser
+sphinx_rtd_theme
+myst_nb
