+++ docs/source/API.rst
@@ -1,0 +1,10 @@
+API
+====================
+
+pyEulerCurves module
+---------------------------------------
+
+.. automodule:: pyEulerCurves
+   :members:
+   :undoc-members:
+   :show-inheritance:
