+++ docs/source/citation.rst
@@ -1,0 +1,19 @@
+Citation
+==================
+If you find this package useful, please consider citing our work:
+
+.. code-block:: bibtex
+
+   @article{10.1093/gigascience/giad094,
+       author = {Dłotko, Paweł and Gurnari, Davide},
+       title = {Euler characteristic curves and profiles: a stable shape invariant for big data problems},
+       journal = {GigaScience},
+       volume = {12},
+       pages = {giad094},
+       year = {2023},
+       month = {11},
+       issn = {2047-217X},
+       doi = {10.1093/gigascience/giad094},
+       url = {https://doi.org/10.1093/gigascience/giad094},
+       eprint = {https://academic.oup.com/gigascience/article-pdf/doi/10.1093/gigascience/giad094/60704121/giad094.pdf},
+   }
