+++ docs/source/examples/index.rst
@@ -1,0 +1,8 @@
+Notebooks
+=========
+
+.. toctree::
+   :maxdepth: 1
+
+   example_pointclouds_VR.ipynb
+   example_cubical.ipynb
