+++ docs/source/index.rst
@@ -1,0 +1,42 @@
+Welcome to pyEulerCurves's documentation!
+==========================================
+
+**pyEulerCurves** is a Python library for parallel computation of Euler Characteristic Curves (ECC) from point cloud and bitmap data.
+
+It provides fast and scalable methods for computing topological signatures useful in shape analysis, machine learning, and topological data analysis.
+
+For theoretical background and applications, refer to the paper:  
+*"Euler characteristic curves and profiles: a stable shape invariant for big data problems"*,  
+`DOI: 10.1093/gigascience/giad094 <https://doi.org/10.1093/gigascience/giad094>`_.
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: 📦 Getting Started
+
+   getting_started
+
+.. toctree::
+   :maxdepth: 2
+   :caption: 🧪 Examples
+
+   examples/index
+
+.. toctree::
+   :maxdepth: 2
+   :caption: 🛠️ API Reference
+
+   API
+
+.. toctree::
+   :maxdepth: 1
+   :caption: 📚 Resources
+
+   citation
+
+
+Indices and tables
+==================
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
