--- examples/data/00000.png_t_u8_50x200x100.from_stack.raw.euler
@@ -1,255 +1,0 @@
-0 3674
-1 7079
-2 10076
-3 12666
-4 15054
-5 17100
-6 18825
-7 20268
-8 21434
-9 22301
-10 22910
-11 23280
-12 23359
-13 23298
-14 23018
-15 22504
-16 21918
-17 21141
-18 20102
-19 18845
-20 17583
-21 16067
-22 14443
-23 12867
-24 11105
-25 9068
-26 7053
-27 4986
-28 2805
-29 551
-30 -1806
-31 -4290
-32 -6791
-33 -9166
-34 -11708
-35 -14332
-36 -16958
-37 -19668
-38 -22289
-39 -24903
-40 -27595
-41 -30396
-42 -32981
-43 -35671
-44 -38389
-45 -41179
-46 -43771
-47 -46481
-48 -49108
-49 -51712
-50 -54318
-51 -56777
-52 -59355
-53 -61706
-54 -64046
-55 -66407
-56 -68687
-57 -71021
-58 -73252
-59 -75393
-60 -77484
-61 -79554
-62 -81475
-63 -83385
-64 -85200
-65 -87105
-66 -88760
-67 -90530
-68 -92034
-69 -93495
-70 -94996
-71 -96354
-72 -97643
-73 -98946
-74 -100127
-75 -101338
-76 -102425
-77 -103438
-78 -104415
-79 -105135
-80 -105927
-81 -106550
-82 -107119
-83 -107545
-84 -108002
-85 -108392
-86 -108696
-87 -109029
-88 -109140
-89 -109332
-91 -109165
-92 -109015
-93 -108792
-94 -108488
-95 -108066
-96 -107681
-97 -107229
-98 -106654
-99 -106100
-100 -105469
-101 -104736
-102 -103859
-103 -103010
-104 -102094
-105 -101040
-106 -99937
-107 -98812
-108 -97654
-109 -96499
-110 -95365
-111 -94042
-112 -92656
-113 -91086
-114 -89527
-115 -87938
-116 -86327
-117 -84703
-118 -82922
-119 -81161
-120 -79411
-121 -77508
-122 -75704
-123 -73838
-124 -72053
-125 -70125
-126 -68199
-127 -66160
-128 -64166
-129 -62105
-130 -59967
-131 -57823
-132 -55698
-133 -53462
-134 -51262
-135 -48950
-136 -46637
-137 -44323
-138 -41941
-139 -39679
-140 -37330
-141 -34948
-142 -32418
-143 -30212
-144 -27845
-145 -25457
-146 -23076
-147 -20735
-148 -18362
-149 -16003
-150 -13663
-151 -11323
-152 -8896
-153 -6467
-154 -4123
-155 -1680
-156 566
-157 2913
-158 5167
-159 7512
-160 9888
-161 12258
-162 14495
-163 16903
-164 19129
-165 21339
-166 23536
-167 25668
-168 27761
-169 29908
-170 32015
-171 34135
-172 36128
-173 38171
-174 40101
-175 41983
-176 43868
-177 45804
-178 47635
-179 49407
-180 51177
-181 52885
-182 54571
-183 56047
-184 57633
-185 59153
-186 60625
-187 62111
-188 63508
-189 64854
-190 66200
-191 67438
-192 68708
-193 69825
-194 70834
-195 71820
-196 72764
-197 73626
-198 74414
-199 75137
-200 75876
-201 76452
-202 77004
-203 77483
-204 77928
-205 78337
-206 78662
-207 79007
-208 79258
-209 79370
-210 79424
-211 79368
-212 79174
-213 78914
-214 78695
-215 78372
-216 77961
-217 77468
-218 76970
-219 76337
-220 75668
-221 74941
-222 74112
-223 73128
-224 72043
-225 70915
-226 69730
-227 68490
-228 67113
-229 65678
-230 64115
-231 62574
-232 60885
-233 59172
-234 57480
-235 55530
-236 53570
-237 51412
-238 49280
-239 47004
-240 44732
-241 42430
-242 39898
-243 37333
-244 34617
-245 31891
-246 29110
-247 26207
-248 23285
-249 20248
-250 17070
-251 13842
-252 10532
-253 7200
-254 3665
-255 1
