--- examples/data/Planck_t_u8_500x1000.raw.euler
@@ -1,164 +1,0 @@
-5 5
-6 8
-7 17
-8 30
-9 57
-10 90
-11 129
-12 177
-13 251
-14 350
-15 432
-16 556
-17 697
-18 868
-19 1053
-20 1261
-21 1442
-22 1671
-23 1942
-24 2229
-25 2611
-26 2921
-27 3181
-28 3541
-29 3894
-30 4245
-31 4656
-32 4974
-33 5333
-34 5710
-35 5955
-36 6331
-37 6668
-38 6861
-39 7158
-40 7350
-41 7500
-42 7570
-43 7710
-44 7783
-45 7773
-46 7760
-47 7681
-48 7568
-49 7354
-50 7167
-51 6892
-52 6633
-53 6347
-54 6056
-55 5651
-56 5163
-57 4566
-58 4151
-59 3684
-60 3094
-61 2518
-62 1961
-63 1298
-64 625
-65 -119
-66 -706
-67 -1163
-68 -1784
-69 -2287
-70 -2822
-71 -3460
-72 -3885
-73 -4557
-74 -5006
-75 -5528
-76 -6132
-77 -6572
-78 -6988
-79 -7374
-80 -7720
-81 -8070
-82 -8312
-83 -8627
-84 -8877
-85 -9118
-86 -9366
-87 -9569
-88 -9772
-89 -9866
-90 -9986
-91 -10059
-92 -9940
-93 -10023
-94 -9970
-95 -9883
-96 -9759
-97 -9649
-98 -9490
-99 -9363
-100 -9257
-101 -9138
-102 -8980
-103 -8691
-104 -8523
-105 -8367
-106 -8150
-107 -7941
-108 -7741
-109 -7430
-110 -7198
-111 -6934
-112 -6704
-113 -6500
-114 -6198
-115 -5913
-116 -5670
-117 -5417
-118 -5147
-119 -4860
-120 -4683
-121 -4390
-122 -4178
-123 -3924
-124 -3676
-125 -3433
-126 -3144
-127 -2957
-128 -2741
-129 -2550
-130 -2379
-131 -2162
-132 -2005
-133 -1842
-134 -1722
-135 -1555
-136 -1405
-137 -1292
-138 -1164
-139 -1070
-140 -925
-141 -844
-142 -753
-143 -688
-144 -622
-145 -515
-146 -447
-147 -396
-148 -338
-149 -293
-150 -239
-151 -201
-152 -171
-153 -139
-154 -119
-155 -91
-156 -61
-157 -47
-158 -33
-159 -29
-160 -25
-161 -17
-162 -15
-163 -13
-164 -11
-165 -7
-166 -1
-167 0
-171 1
