+++ pyEulerCurves/_version.py
@@ -1,2 +1,2 @@
- # -*- coding: utf-8 -*-
-__version__ = "0.1"
+# -*- coding: utf-8 -*-
+__version__ = "0.3"
