+++ setup.py
@@ -1,20 +1,20 @@
 from setuptools import setup, find_packages
 from pybind11.setup_helpers import Pybind11Extension, build_ext
 
-__version__ = 0.3
+__version__ = "0.3"
 
 ext_modules = [
     Pybind11Extension(
         "pyEulerCurves._compute_local_EC_cubical",
-        ["src/compute_local_EC_cubical.cpp"],
+        ["pyEulerCurves/src/compute_local_EC_cubical.cpp"],
         define_macros=[("VERSION_INFO", __version__)],
-        include_dirs=["src"],
+        include_dirs=["pyEulerCurves/src"],
     ),
     Pybind11Extension(
         "pyEulerCurves._compute_local_EC_VR",
-        ["src/compute_local_EC_VR.cpp"],
+        ["pyEulerCurves/src/compute_local_EC_VR.cpp"],
         define_macros=[("VERSION_INFO", __version__)],
-        include_dirs=["src"],
+        include_dirs=["pyEulerCurves/src"],
     ),
 ]
 
@@ -34,8 +34,15 @@ setup(
     package_data={
         "pyEulerCurves": [
             "*.py",
-            "../src/*.h",
-            "../src/*.cpp",
+            "src/*.h",
+            "src/*.cpp",
         ],  # Include header and source files
     },
+    install_requires=[
+        "numpy",
+        "matplotlib",
+        "scikit-learn",
+        "pybind11>=2.10.0",
+        "gudhi",
+    ],
 )
